[UPD] Badly implemented gravity
Gravity but buggy
Attempted to separate Engine calculation into a function that requires the model update, and a function that does not. Both functions are selectively run by a setInterval locked to roughly 30fps. The non-model-requiring function does not run independently from the other function [bug] when it should run every 30ms, which would guarantee gravity to work. No idea why